reorganize folders

+++ sd_fused/utils/tensors/softmax.py
@@ -7,12 +7,6 @@ def softmax(x: Tensor, *, dim: int) -> Tensor:
     """In-place softmax implementation."""
 
     dtype = x.dtype
-
-    # TODO test only half!
-    # TODO use in-place?
-    # x = x.float()
-    x = x.softmax(dim)
-    # x = x.softmax(dim, dtype=torch.float32) # ? is this better?
-    # x = x.to(dtype)
+    x = x.softmax(dim, dtype=dtype)
 
     return x
